New FC_ObjectTasker library (#5)
New library flies after tests and comparison to the function pointers Tasker. Results are quite the same but this solution is a way more powerful (can fully replace lambda expressions which forced to use function objects with a large in size and slow on Arduino c++11 standard library).

Also that new ObjectTasker is easier in expanding its capabilities by overriding some Tasker methods.