Easier way of adding new tasks to the FC_ObjectTasker (#7)
Now there is no need to overload the FC_Task constructor to set up interval and maxDuration. FC_Task only override the execute() method and implement there the actual task code. Interval and maxDuration are set when adding new Task to the FC_ObjectTasker instance. Inside that function this values are set (because Tasker is friend class of Task and can use private Task setProperties() method)